tests/wpt/web-platform-tests/css/css-tables/inheritance.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>Inheritance of CSS Tables properties</title>
<link rel="help" href="https://drafts.csswg.org/css-tables-3/#property-index">
<meta name="assert" content="Properties inherit or not according to the spec.">
<meta name="assert" content="Properties have initial values according to the spec.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/inheritance-testcommon.js"></script>
</head>
<body>
<div id="container">
  <div id="target"></div>
</div>
<script>
assert_inherited('border-collapse', 'separate', 'collapse');
assert_inherited('border-spacing', '0px 0px', '10px 20px');
assert_inherited('caption-side', 'top', 'bottom');
assert_inherited('empty-cells', 'show', 'hide');
assert_not_inherited('table-layout', 'auto', 'fixed');
</script>
</body>
</html>
