third_party/blink/web_tests/external/wpt/css/css-tables/inheritance-expected.txt

This is a testharness.js-based test.
[FAIL] Property border-spacing has initial value 0px
  assert_equals: expected "0px" but got "0px 0px"
Harness: the test ran to completion.
